Move cogs out of subfolders
